moved payload creation to different method and file
